fix: apply hotfix to support Android 11
The current version of the plugin is not able to open databases when targeting the SDK 30
and running on devices with Android 11.
The issue was already reported https://github.com/storesafe/cordova-sqlite-storage-dependencies/issues/3
and https://github.com/storesafe/cordova-sqlite-storage/issues/954
From what I was able to understand, it is related with the permissions changes on Android 11,
where the SQLite is unable to open the database file using native C primitive `sqlite3_open_v2`:
https://github.com/liteglue/Android-sqlite-connector/blob/master/src/io/liteglue/SQLGDatabaseHandle.java#L14
https://github.com/liteglue/Android-sqlite-connector/blob/master/src/io/liteglue/SQLiteNative.java#L60
https://github.com/liteglue/Android-sqlite-native-driver/blob/master/native/sqlc.c#L40
causing the handler to crash with a NullPointerException when throwing an exception
https://github.com/liteglue/Android-sqlite-connector/blob/master/src/io/liteglue/SQLiteGlueConnection.java#L12

Summing up, this hotfix will enable the usage of the default Android database provider for Android 11 only,
by setting the `oldImpl` to `true`, which is the same behaviour of using the `androidDatabaseProvider: 'system'`
setting at the javascript side.

References https://outsystemsrd.atlassian.net/browse/RNMT-4515